+++ js/functions.js
@@ -39,3 +39,12 @@ function checkNumber(string){                                  // 3 задача
 checkNumber('2023 год');
 checkNumber(('а я томат'));
 
+const getRandomNumber = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+export{getRandomNumber};
+
